Move scope to TypeSpecs
